# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Pokémon Shiny Encounter Counter — REST API",
        "title": "Encounty API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/pokedex": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pokedex"
                ],
                "summary": "Get the Pokedex",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_server.PokedexEntry"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errResp"
                        }
                    }
                }
            }
        },
...
        "/sync/pokemon": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pokedex"
                ],
                "summary": "Sync Pokedex from PokeAPI",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server.PokedexSyncResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_server.errResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_server.PokedexEntry": {
            "type": "object",
            "properties": {
                "canonical": {
                    "type": "string"
                },
                "forms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_server.PokemonForm"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "names": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "internal_server.PokedexSyncResponse": {
            "type": "object",
            "properties": {
                "added": {
                    "type": "integer"
                },
                "namesUpdated": {
                    "type": "integer"
                },
                "new": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_server.PokemonForm": {
            "type": "object",
            "properties": {
                "canonical": {
                    "type": "string"
                },
                "names": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "sprite_id": {
                    "type": "integer"
                }
            }
        },
...
        "internal_server.errResp": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}